+++ OnboardPietroPieri/components/header.tsx
@@ -1,11 +1,17 @@
 import React from "react";
-import { HeaderContainer, HeaderText } from "../styles/styles";
-import { HeaderProps } from "../interfaces/component";
+import { HeaderContainer, HeaderText, Spacing } from "../styles/styles";
+
+export interface HeaderProps {
+  text: string;
+}
 
 export const Header: React.FC<HeaderProps> = ({ text }) => {
   return (
-    <HeaderContainer>
-      <HeaderText>{text}</HeaderText>
-    </HeaderContainer>
+    <>
+      <HeaderContainer>
+        <HeaderText>{text}</HeaderText>
+      </HeaderContainer>
+      <Spacing />
+    </>
   );
 };
